refactor: split onboard review render
